+++ stream-chat-android-client/src/main/java/io/getstream/chat/android/client/api2/mapping/DomainMapping.kt
@@ -246,6 +246,7 @@ internal class DomainMapping(
                 reminder = reminder?.toDomain(),
                 sharedLocation = shared_location?.toDomain(),
                 channelRole = member?.channel_role,
+                deletedForMe = deleted_for_me ?: false,
                 extraData = extraData.toMutableMap(),
             ).let(messageTransformer::transform)
         }