+++ stream-chat-android-client/src/main/java/io/getstream/chat/android/client/api2/mapping/EventMapping.kt
@@ -403,6 +403,7 @@ internal class EventMapping(
             message = message.toDomain(),
             hardDelete = hard_delete ?: false,
             channelMessageCount = channel_message_count,
+            deletedForMe = deleted_for_me ?: false,
         )
     }